+++ lib/posix/posix_utils.nim
@@ -99,3 +99,11 @@ proc mkdtemp*(prefix: string): string =
     raise newException(OSError, $strerror(errno))
   return $tmpl
 
+when defined(linux):
+  proc isSsd*(diskLetter: char): bool {.inline.} =
+    ## Returns ``true`` if disk is SSD (Solid). Linux only.
+    ##
+    ## .. code-block:: nim
+    ##   echo isSsd('a') ## 'a' for /dev/sda, 'b' for /dev/sdb, ...
+    ##
+    try: readFile("/sys/block/sd" & diskLetter & "/queue/rotational") == "0\n" except: false